requirements.txt
openai>=1.2
streamlit-feedback
streamlit-extras
streamlit-card
streamlit-audiorec
pandas